% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{asesgeo}
\alias{asesgeo}
\alias{asesgeo-package}
\title{asesgeo: An R Geographic Analytics Toolkit for ASES}
\description{
An analytic toolkit comprising of a series of work functions on geographics,
e.g., coordinate conversion for Chinese coordinates/addresses, IP parsing, ...
}
\details{
This package is comprised of \cr \describe{
  \item{Coordinate conversion}{\itemize{
    \item Local tranformer: \code{\link{wgs2gcj}()}, \code{\link{wgs2bd}()}, 
      \code{\link{gcj2wgs}()}, \code{\link{gcj2bd}()}, \code{\link{bd2wgs}()}, 
      \code{\link{bd2gcj}()}
    \item Generic wrapper: \code{\link{conv_coord}()}
  }}
  \item{Geohost (ip parser)}{\code{\link{geohost}()} or \code{\link{parse_geohost}()}}
  \item{Geocode (address locator)}{\code{\link{geocode}()} or \code{\link{parse_geocode}()}}
  \item{Reverse geocode (coordinate locator)}{\code{\link{revgeocode}()} or 
    \code{\link{parse_revgeocode}()}}
  \item{Helpers}{\itemize{
    \item Formatter: \code{\link{coord_format}}, \code{\link{lat_coord}()}, 
      \code{\link{lon_coord}()}
    \item Inside or outside China: \code{\link{is_out_of_china}()}
    \item Show my IP: \code{\link{show_my_ip}()}     
  }}
}
}
\seealso{
\pkg{\link{aseskit}}
}
\author{
\strong{Creator, Maintainer}: Yiying Wang, \email{wangy@aetna.com}
}
\keyword{internal}
