% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohost.R
\name{geohost}
\alias{geohost}
\title{IP address lookup}
\usage{
geohost(ip = "", api = c("ipinfo", "ipify", "ipstack"),
  output = c("country", "region", "city", "all", "raw"),
  use_curl = FALSE, time = 0, key = NULL)
}
\arguments{
\item{ip}{a character vector specifying IPs (IPv4 or IPv6) (e.g., "12.215.42.19").
The default value is no IP is specified and the host IP is used.}

\item{api}{use 'ipify' (ipify.org), 'ipstack' (ipstack.com) or 'ipinfo'(ipinfo.io) 
lookup API. By default 'ipinfo' is used.}

\item{output}{character, 'country', 'region', 'city', 'all' or 'raw'. It decides
which columns to return in the output data.frame. \itemize{
 \item 'country': ip and country \cr
 \item 'region': ip, region and country \cr
 \item 'city': ip, city, region and country \cr
 \item 'all': all the info \cr
 \item 'raw': the raw JSON list. You can then apply \code{\link{parse_geohosts}}
 to process the list for structured data.
}}

\item{use_curl}{logical, whether use curl to access the APIs. Default FALSE.}

\item{time}{numeric, time interval to lookup the IP address, by seconds. It is
used to avoid overuse of the APIs. Default 0.}

\item{key}{character, the api key for ipinfo or ipstack. Default NULL, which
indicates that the function will try to find it in cache. When no match is 
found, a GUI wizard will be launched for you to enter the key. If the API does
not call for a key, set it to NA.}
}
\value{
a data.frame.
}
\description{
geocodes an IP address using either ipify.org (\url{https://www.ipify.org}), 
ipstack.com (\url{https://ipstack.com}) or ipinfo.io (\url{http://ipinfo.io/developers}) 
IP lookup API.
}
\details{
note that ipstack.com API is allowed up to 10,000 queries per month
by default, ipinfo API is limited to 1,000 requests per day, and ipify.org API
is limited to 1,000 requests per month.
}
\examples{
\dontrun{
# geocode host IP
geohost()
geohost(api = "ipinfo")

# specify an IP for geocoding
geohost(ip = "12.215.42.19")
geohost(ip = "12.215.42.19", api = "ipinfo")

# geocode multiple IPs
geohost(ip = c("61.135.169.81", "12.215.42.19"))
geohost(ip = c("61.135.169.81", "12.215.42.19"), api = "ipinfo")
}
}
\references{
\itemize{
 \item ipipy.org IP lookup API at \url{https://www.ipify.org} \cr
 \item ipstack.com IP lookup API at \url{https://ipstack.com/documentation} \cr
 \item ipinfo.io IP lookup API at \url{https://ipinfo.io/developers}
}
}
\seealso{
\code{\link{geocode}}, \code{\link{revgeocode}}, \code{\link{parse_geohosts}}.
}
\author{
Jun Cai (\email{cai-j12@mails.tsinghua.edu.cn}), PhD student from
Center for Earth System Science, Tsinghua University \cr
Yiying Wang (\email{wangy@aetna.com})
}
