% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isOutOfChina}
\alias{isOutOfChina}
\alias{is_out_of_china}
\title{Detect whether the points are outside China}
\usage{
isOutOfChina(y, ..., accurate = TRUE, gcs = c("WGS-84", "GCJ-02",
  "BD-09"), plot = FALSE, canvas = c("china", "world"),
  xlim = c(72.004, 137.8347), ylim = c(0.8293, 55.8271))
}
\arguments{
\item{y}{one of \itemize{
\item coordinate pair: a list of coordinate vectors \code{list(lat1, lon1)},
in which case you can provide other coordiante vectors in \code{...} \cr
\item latitude vector: latitude number(s), in which case you need to provice 
 longitute number(s) in \code{...} \cr
\item coordinate matrix: a matrix (row 1-2 or col 1-2). The function will 
 choose how to read the data \cr
\item coordinate data.frame: a data.frame (col 1-2)
}}

\item{...}{one of \itemize{
\item coordinate pairs: when \code{y} is a list containing a coordinate pair 
 (e.g., \code{list(lat1, lon1)}), you can pass other coordinate pairs here 
 (e.g., \code{list(lat2, lon2), list(lat3, lon3), ...}). \cr
\item longitude vector: when \code{y} is only latitude vector, you can pass 
 \code{x} (longitude) here. \cr
\item when \code{y} is a coordinate matrix or data.frame, \code{...} is omitted.
}}

\item{accurate}{logical, whether use accurate China boundary to analyze
the coordinates. Default TRUE. When using the accurate China boundary,
the function will check if the coordinates (Lat, Lon) are actually inside 
China's territory. When set FALSE, the function will use a very vague rectangle 
defined by \code{xlim} and \code{ylim} to reprensent China boundary.}

\item{gcs}{character, the geo-coordinate system, 'WGS-84', 'GCJ-02' or 'BD-09'.
Default 'WGS-84'. The function will analyze the points (Lat, Lon) under the
specific gcs.}

\item{plot}{logical, if generate a plot to show the location of the points.}

\item{canvas}{character, 'china' or 'world'. It is only effective when 
\code{plot} == TRUE. If 'china', the plot is restricted to east Asia, with
the projection 'azequalarea', orientation of c(30, 105) with no rotation. If
'world', the plot is the full 'world' map, with the projetion 'mercator'.}

\item{xlim}{numeric vector of length 2. The longitude limits of the China map.
Default c(72.004, 137.8347). when \code{accurate} is FALSE, the rectangle
made by xlim and ylim will be used as China.}

\item{ylim}{numeric vector of length 2. The latitude limits of the China map.
Default c(0.8293, 55.8271). when \code{accurate} is FALSE, the rectangle
made by xlim and ylim will be used as China boundary.}
}
\value{
depends on \code{plot}, \itemize{
 \item \code{TRUE}: returns a 'ggplot2' object. \cr
 \item \code{FALSE}; returns a logical vector with the same length of Lat (or Lon).
}
}
\description{
Analyzes the coordinates (lat, lon) to check whether they are in or out of China. 
You can choose to return a logical vector or a ggplot2 object for quick 
visualization.
}
\details{
The most popular open-source coordinate transformation algorithm 
regards China as a rectangle. Points outside this rectangle are treated as 
'outside China', of which case the conversion from WGS-84 ==> GCJ-02 / BD-09
is not conducted. This function inherits this algorithm as well. As a step 
further, it prepares an 'accurate' method that use the actual China boundary
to analyze the input points. Due to the limitation of proj.4 library, it plots
the points and China boundary under 'WGS-84' system.
}
\examples{
\dontrun{
is_out_of_china(c(10, 30, 40, 50, 70), c(100, 120, 100, 140, 160), 
                accurate_=TRUE)  
# 10.00,100.00 30.00,120.00 40.00,100.00 50.00,140.00 70.00,160.00 
# TRUE         FALSE        FALSE        TRUE         TRUE

is_out_of_china(c(10, 30, 40, 50, 70), c(100, 120, 100, 140, 160), 
                accurate=FALSE)
# 10.00,100.00 30.00,120.00 40.00,100.00 50.00,140.00 70.00,160.00 
# FALSE        FALSE        FALSE        TRUE         TRUE 

## visualize the positions
is_out_of_china(c(10, 30, 40, 50, 70), c(100, 120, 100, 140, 160), 
                plot=TRUE, canvas='china')
is_out_of_china(c(10, 30, 40, 50, 70), c(100, 120, 100, 140, 160), 
                plot=TRUE, canvas='world')              
}
}
\seealso{
Refer to \code{\link{transform_coord}} function family for argument \code{y, ...}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
