% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohost.R
\name{parse_geohosts}
\alias{parse_geohosts}
\alias{parse_geohost}
\title{Parse geohost API results}
\usage{
parse_geohosts(ghlst, output = c("country", "region", "city", "all"),
  ...)
}
\arguments{
\item{ghlst}{a list comprising of lists of class 'ipify_geohost', 'ipinfo_geohost'
or 'ipstack_geohost'.}

\item{output}{character, 'country', 'region', 'city', 'all', or 'raw'. Refer to
\code{\link{geohost}}.}

\item{...}{other arguments to pass to the function}
}
\value{
a \code{\link[tibble]{tibble}}
}
\description{
This is a wrapper of API result parsers for ipify, ipinfo and ipstack. 
If \code{\link{geohost}} yielded 'raw' results, you can then convert the 
output lists to a structured data using this function.
}
\examples{
\dontrun{
ghlst <- geohost(c('<IP1>', '<IP2>'), output='raw')  # yield a list
parse_geohosts(ghlst) 
}
}
\seealso{
\code{\link{parse_api_data}}, \code{\link{geohost}}, 
\code{\link{geocode}}, \code{\link{parse_geocode}};
\code{\link{revgeocode}}, \code{\link{parse_revgeocode}}
}
