% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeocode.R
\name{parse_revgeocodes}
\alias{parse_revgeocodes}
\alias{parse_revgeocode}
\title{Parse revgeocode API results}
\usage{
parse_revgeocodes(rgclst, output = c("address", "addressc", "all"),
  gcs = NULL, idf = TRUE, ...)
}
\arguments{
\item{rgclst}{a list comprising of lists of class 'google_revgeocode', 
'baidu_revgeocode' or 'gaode_revgeocode', subclasses of 'api_data'.}

\item{output}{character, 'address', 'addressc', or 'all'. Default 
'address'. Refer to \code{\link{revgeocode}()}.}

\item{gcs}{character, the geo-coordinate system for output. 'WGS-84', 'GCJ-02', 
or 'BD-09'. Default NULL, which indicates that no coordinate conversion will be 
conducted on the results.}

\item{idf}{logical, whether generate an identifier column 'idf' in the output. 
Default TRUE.}

\item{...}{optional arguments. Generic args: \itemize{
 \item ics_china: default coordinate system for locations in China, 'WGS-84', 
  'GCJ-02', or 'BD-09' \cr
 \item ics_intl: default coordinate system for locations outside China, 'WGS-84', 
  'GCJ-02', or 'BD-09'  
}
API-specific args: \describe{
 \item{google api}{\itemize{
  \item name_type: character 'long' or 'short', indicating whether to extract
   long_name or short_name. Default 'long'.
 }}
 \item{baidu api}{\itemize{
  \item (not used yet)
 }}
 \item{gaode api}{\itemize{
  \item (not used yet)
 }}
}}
}
\value{
a \code{\link[tibble]{tibble}}
}
\description{
This is a wrapper of API result parsers for google, baidu and gaode map. 
It can serve as a subsequent handler to convert 'raw' results yielded
using \code{\link{revgeocode}()} to a structured data.frame.
}
\examples{
\dontrun{
rgc <- revgeocode(c(<lat1>, <lon1>), api='google', output='raw')
parse_revgeocodes(rgc, output='all')
}
}
\seealso{
\code{\link{parse_api_data}()}, \code{\link{revgeocode}()},
\code{\link{geocode}()}, \code{\link{parse_geocodes}()};
\code{\link{geohost}()}, \code{\link{parse_geohosts}()};
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
