% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geohost.R
\name{show_my_ip}
\alias{show_my_ip}
\title{Display my own IP address}
\usage{
show_my_ip()
}
\value{
character, current IP address. If fails, returns NULL.
}
\description{
Call ipify.org API to detect my own IP address. If you cannot get access to
the API, NULL will be returned.
}
\examples{
\dontrun{
show_my_ip()
}
}
