% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{synthesize_baidumap_api}
\alias{synthesize_baidumap_api}
\alias{synthesize_baidu_api_geocode}
\alias{synthesize_baidu_api_revgeocode}
\alias{synthesize_baidu_api_convcoord}
\title{Synthesize baidumap API}
\usage{
synthesize_baidumap_api(url_body, api = c("geocode", "revgeocode",
  "convcoord"), use_curl = FALSE, ...)

synthesize_baidu_api_geocode(url_body, ics = c("BD-09", "GCJ-02",
  "WGS-84"), ocs = c("BD-09", "GCJ-02", "WGS-84"), city = NULL,
  use_curl = FALSE, ...)

synthesize_baidu_api_revgeocode(url_body, ics = c("BD-09", "GCJ-02",
  "WGS-84"), ocs = c("BD-09", "GCJ-02", "WGS-84"), pois = NULL,
  radius = NULL, extensions_road = FALSE, extensions_town = FALSE,
  language = NULL, language_auto = NULL, latest_admin = 0,
  use_curl = FALSE, ...)

synthesize_baidu_api_convcoord(url_body, coord_from = c("BD-09",
  "GCJ-02", "WGS-84"), coord_to = c("BD-09", "GCJ-02", "WGS-84"),
  use_curl = FALSE, ...)
}
\arguments{
\item{url_body}{character, the leading argument in the API URL.}

\item{api}{character, 'geocode', 'revgeocode', 'convcoord'.}

\item{use_curl}{logical, whether use \code{curl}(). Default TRUE.}

\item{...}{other arguments to pass to the function, e.g., \code{key}}

\item{ics}{character, for geocoding. 'BD-09', 'GCJ-02' or 'WGS-84'.}

\item{ocs}{character, for geocoding. 'BD-09', 'GCJ-02' or 'WGS-84'}

\item{city}{character, for geocoding. A valid Chinese city name to limit the search range.}

\item{pois}{numeric 1 or 0, for revgeocode. whether callback surrounding pois
(1=Yes, 0=No). You will need to apply for special service access when calling 
pois outside China. Default NULL (0).}

\item{radius}{numeric 0-1000, for revgeocode. The radius (m) for calling back pois.
Default NULL (indicating 1000).}

\item{extensions_road}{logical, for revgeocode. Callback 3 streets near the spot 
when set to \code{TRUE}. Default FALSE.}

\item{extensions_town}{logical, for revgeocode. Callback info of town when set 
to \code{TRUE}. Default FALSE.}

\item{language}{character, for revgeocode. Valid language: "local" or "el", "gu", 
"en", "vi", "ca", "it", "iw", "sv", "eu", "ar", "cs", "gl", "id", "es", "ru", 
"sr", "nl", "pt", "tr", "tl", "lv", "lt", "th", "ro", "fil", "ta", "fr", "bg", 
"hr", "bn", "de", "hu", "fa", "hi", "fi", "da","ja", "te", "ml", "ko", "kn", 
"sk", "pl", "uk", "sl", "mr", "local", "en-GB", "en-AU", "zh-TW", "pt-BR", 
"pt-PT", "zh-CN". Default NULL.}

\item{language_auto}{numeric 1 or 0, for revgeocode. whether auto fill the 
admin area (1=Yes, 0=No). You will need to apply for special service access
when using multi-language functionality. Default NULL (ignored).}

\item{latest_admin}{numeric 1 or 0, for revgeocode. whether get the latest 
administrative area info (1=Yes, 0=No). Default 0.}

\item{coord_from}{character, for conv coord. 'BD-09', 'GCJ-02', or 'WGS-84'}

\item{coord_to}{character, for conv coord. 'BD-09', 'GCJ-02', or 'WGS-84'}
}
\value{
A \code{url} object for \code{\link{get_api_data}()} use.
}
\description{
Working function for \code{\link[aseskit]{synthesize_api}()}. It will call
lower functions \code{\link{synthesize_baidu_api_geocode}()} or 
\code{\link{synthesize_baidu_api_revgeocode}()}
or \code{synthesize_baidu_api_convcoord}.
}
\note{
When \code{key} is omitted or set NULL, the function will search for it
in the cache. If nothing is found, a GUI wizard will be launched for you to 
input the API key. If you want to bypass API key, set \code{key} = NA.
}
\examples{
\dontrun{
synthesize_baidumap_api("<Address>", key="<YOUR BAIDUMAP KEY>")
}
}
\references{
Baidu map API documentation: 
\url{http://lbsyun.baidu.com/index.php?title=webapi/guide}
}
