% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{synthesize_gaodemap_api}
\alias{synthesize_gaodemap_api}
\alias{synthesize_gaode_api_geocode}
\alias{synthesize_gaode_api_revgeocode}
\alias{synthesize_gaode_api_convcoord}
\title{Synthesize gaode map API}
\usage{
synthesize_gaodemap_api(url_body, api = c("geocode", "revgeocode",
  "convcoord"), use_curl = FALSE, ...)

synthesize_gaode_api_geocode(url_body, batch = FALSE, city = NULL,
  use_curl = FALSE, ...)

synthesize_gaode_api_revgeocode(url_body, batch = FALSE,
  extensions = c("all", "base"), poitype = NULL, radius = NULL,
  roadlevel = NULL, homeorcorp = NULL, use_curl = FALSE, ...)

synthesize_gaode_api_convcoord(url_body, batch = FALSE,
  coord_from = c("WGS-84", "BD-09", "GCJ-02"), coord_to = c("GCJ-02",
  "BD-09", "WGS-84"), use_curl = FALSE, ...)
}
\arguments{
\item{url_body}{character, the leading argument in the API URL. If you want to
use batch mode, you can synthesize them with pipe "|" (no more than 10 per batch).
E.g., "beijing|tianjin|shanghai".}

\item{api}{character, 'geocode', 'revgeocode', 'convcoord'.}

\item{use_curl}{logical, whether use \code{curl}(). Default TRUE.}

\item{...}{other arguments to pass to the function, e.g., \code{key}}

\item{batch}{logical, for geocoding, If use batch inquery mode. Default FALSE.}

\item{city}{character, for geocoding. A valid Chinese city name to limit results.}

\item{extensions}{character, 'all' or 'base', for revgeocode. Return basic 
info ('base') or base+additional info ('all'). Default "all".}

\item{poitype}{character, for revgeocode. One or more POI type codes (separated 
by a pipe '|'). The valid POI type codes refers to 
\url{https://lbs.amap.com/api/webservice/download}. 
It will not take effect when applying batch mode (multiple \code{latlng}s 
are provided). Only effective when \code{extensions=="all"}. Default NULL.}

\item{radius}{numeric (0-3000), for revgeocode. The radius (m) for calling 
back pois. Default NULL (1000).}

\item{roadlevel}{numeric 0 or 1, for revgeocode. Return all the roads (0) or 
main roads only (1)}

\item{homeorcorp}{numeric 0-2, for revgeocode. Order the callback pois, \itemize{
\item 0: do not intervene the order of the pois. \cr
\item 1: home-related pois are prioritized. \cr
\item 2: coorporate-related pois are prioritized.
} Default NULL (0).}

\item{coord_from}{character, for conv coord. 'BD-09', 'GCJ-02', or 'WGS-84'. 
Default 'GCJ-02'.}

\item{coord_to}{character, for conv coord. 'BD-09', 'GCJ-02', or 'WGS-84'.
Only 'GCJ-02' is now supported.}
}
\value{
A \code{url} object for \code{\link{get_api_data}()} use.
}
\description{
Working function for \code{\link[aseskit]{synthesize_api}()}. It will call
lower functions \code{\link{synthesize_gaode_api_geocode}()} or 
\code{\link{synthesize_gaode_api_revgeocode}()}
or \code{\link{synthesize_gaode_api_convcoord}()}.
}
\note{
When \code{key} is omitted or set NULL, the function will search for it
in the cache. If nothing is found, a GUI wizard will be launched for you to 
input the API key. If you want to bypass API key, set \code{key} = NA.
}
\examples{
\dontrun{
synthesize_gaodemap_api("<Address>", key="<YOUR GAODEMAP KEY>")
}
}
\references{
Gaode map API documentation: 
\url{https://lbs.amap.com/api/webservice/guide/api}
}
