% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{synthesize_ipify_api}
\alias{synthesize_ipify_api}
\alias{synthesize_ipify_api_geohost}
\alias{synthesize_ipify_api_getip}
\title{Synthesize ipify API}
\usage{
synthesize_ipify_api(url_body, api = c("geohost", "getip"),
  use_curl = FALSE, ...)

synthesize_ipify_api_geohost(url_body, use_curl = FALSE, ...)

synthesize_ipify_api_getip(url_body, use_curl = FALSE, ...)
}
\arguments{
\item{url_body}{character, the leading argument of the API URL.}

\item{api}{character, 'geohost', 'getip'.}

\item{use_curl}{logical, whether use \code{curl}(). Default TRUE.}

\item{...}{other arguments to pass to the function, e.g., \code{key}}
}
\value{
A \code{url} object for \code{\link{get_api_data}()} use.
}
\description{
Working function for \code{\link[aseskit]{synthesize_api}()}. It will call
lower functions \code{\link{synthesize_ipify_api_geohost}()} or 
\code{\link{synthesize_ipify_api_getip}()}.
}
\note{
When \code{key} is omitted or set NULL, the function will search for it
in the cache. If nothing is found, a GUI wizard will be launched for you to 
input the API key. If you want to bypass API key, set \code{key} = NA.
}
\examples{
\dontrun{
synthesize_ipify_api("<IP Address>", key="<YOUR IPIFY KEY>")
}
}
\references{
ipify API documentation: 
\url{https://www.ipify.org}
}
