% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{synthesize_ipinfo_api}
\alias{synthesize_ipinfo_api}
\alias{synthesize_ipinfo_api_geohost}
\title{Synthesize ipinfo API}
\usage{
synthesize_ipinfo_api(url_body, api = c("geohost"), use_curl = FALSE,
  ...)

synthesize_ipinfo_api_geohost(url_body, use_curl = FALSE, ...)
}
\arguments{
\item{url_body}{character, the leading argument in the API URL.}

\item{api}{character, 'geohost'.}

\item{use_curl}{logical, whether use \code{curl}(). Default TRUE.}

\item{...}{other arguments to pass to the function, e.g., \code{key}}
}
\value{
A \code{url} object for \code{\link{get_api_data}()} use.
}
\description{
Working function for \code{\link[aseskit]{synthesize_api}()}. It will call
lower functions \code{\link{synthesize_ipinfo_api_geohost}()}.
}
\note{
When \code{key} is omitted or set NULL, the function will search for it
in the cache. If nothing is found, a GUI wizard will be launched for you to 
input the API key. If you want to bypass API key, set \code{key} = NA.
}
\examples{
\dontrun{
synthesize_ipinfo_api("<IP Address>", key="<YOUR IPINFO KEY>")
}
}
\references{
ipinfo API documentation: 
\url{https://ipinfo.io/developers}
}
